+++ airflow/providers/amazon/aws/triggers/ecs.py
@@ -165,8 +165,8 @@ class TaskDoneTrigger(BaseTrigger):
             # fmt: on
             waiter = ecs_client.get_waiter("tasks_stopped")
             logs_token = None
-            while self.waiter_max_attempts >= 1:
-                self.waiter_max_attempts = self.waiter_max_attempts - 1
+            while self.waiter_max_attempts:
+                self.waiter_max_attempts -= 1
                 try:
                     await waiter.wait(
                         cluster=self.cluster, tasks=[self.task_arn], WaiterConfig={"MaxAttempts": 1}
@@ -186,7 +186,7 @@ class TaskDoneTrigger(BaseTrigger):
 
     async def _forward_logs(self, logs_client, next_token: str | None = None) -> str | None:
         """
-        Reads logs from the cloudwatch stream and prints them to the task logs.
+        Read logs from the cloudwatch stream and print them to the task logs.
 
         :return: the token to pass to the next iteration to resume where we started.
         """