+++ tests/www/views/test_views_tasks.py
@@ -29,7 +29,7 @@ import time_machine
 
 from airflow import settings
 from airflow.exceptions import AirflowException
-from airflow.models import DAG, DagBag, DagModel, TaskFail, TaskInstance, TaskReschedule
+from airflow.models import DAG, DagBag, DagModel, TaskFail, TaskInstance, TaskReschedule, XCom
 from airflow.models.dagcode import DagCode
 from airflow.operators.bash import BashOperator
 from airflow.providers.celery.executors.celery_executor import CeleryExecutor
@@ -42,7 +42,7 @@ from airflow.utils.types import DagRunType
 from airflow.www.views import TaskInstanceModelView
 from tests.test_utils.api_connexion_utils import create_user, delete_roles, delete_user
 from tests.test_utils.config import conf_vars
-from tests.test_utils.db import clear_db_runs
+from tests.test_utils.db import clear_db_runs, clear_db_xcom
 from tests.test_utils.www import check_content_in_response, check_content_not_in_response, client_with_login
 
 DEFAULT_DATE = timezone.utcnow().replace(hour=0, minute=0, second=0, microsecond=0)
@@ -68,6 +68,13 @@ def init_dagruns(app, reset_dagruns):
             data_interval=(DEFAULT_DATE, DEFAULT_DATE),
             start_date=timezone.utcnow(),
             state=State.RUNNING,
+        )
+        XCom.set(
+            key="return_value",
+            value="{'x':1}",
+            task_id="runme_0",
+            dag_id="example_bash_operator",
+            execution_date=DEFAULT_DATE,
         )
         app.dag_bag.get_dag("example_subdag_operator").create_dagrun(
             run_id=DEFAULT_DAGRUN,
@@ -103,6 +110,7 @@ def init_dagruns(app, reset_dagruns):
         )
     yield
     clear_db_runs()
+    clear_db_xcom()
 
 
 @pytest.fixture(scope="module")
@@ -329,6 +337,17 @@ def test_views_get(admin_client, url, contents):
     resp = admin_client.get(url, follow_redirects=True)
     for content in contents:
         check_content_in_response(content, resp)
+
+
+def test_xcom_return_value_is_not_bytes(admin_client):
+    url = f"xcom?dag_id=example_bash_operator&task_id=runme_0&execution_date={DEFAULT_VAL}&map_index=-1"
+    resp = admin_client.get(url, follow_redirects=True)
+    # check that {"x":1} is in the response
+    content = "{&#39;x&#39;:1}"
+    check_content_in_response(content, resp)
+    # check that b'{"x":1}' is not in the response
+    content = "b&#39;&#34;{\\&#39;x\\&#39;:1}&#34;&#39;"
+    check_content_not_in_response(content, resp)
 
 
 def test_rendered_task_view(admin_client):